flytt komponenter som lå ensom i mappe direkte i komponenter-mappen, og rename samhandler-mappen
