+++ src/frontend/sider/Fagsak/Behandling/Høyremeny/Hendelsesoversikt/BrevModul/Brevskjema.tsx
@@ -36,8 +36,8 @@ import useDokument from '../../../../../../hooks/useDokument';
 import BrevmottakerListe from '../../../../../../komponenter/brevmottaker/BrevmottakerListe';
 import Datovelger from '../../../../../../komponenter/datovelger/Datovelger';
 import Knapperekke from '../../../../../../komponenter/Knapperekke';
-import PdfVisningModal from '../../../../../../komponenter/PdfVisningModal/PdfVisningModal';
-import { useSamhandlerRequest } from '../../../../../../komponenter/Samhandler/useSamhandler';
+import PdfVisningModal from '../../../../../../komponenter/PdfVisningModal';
+import { useSamhandlerRequest } from '../../../../../../komponenter/samhandler/useSamhandler';
 import type { IBehandling } from '../../../../../../typer/behandling';
 import type { IManueltBrevRequestPåBehandling } from '../../../../../../typer/dokument';
 import type { IPersonInfo } from '../../../../../../typer/person';
