+++ src/frontend/sider/ManuellJournalfør/BrukerPanel.tsx
@@ -20,8 +20,8 @@ import { RessursStatus } from '@navikt/familie-typer';
 import { DeltagerInfo } from './DeltagerInfo';
 import { useManuellJournalfør } from '../../context/ManuellJournalførContext';
 import { KontoSirkel } from '../../ikoner/KontoSirkel';
-import { SamhandlerTabell } from '../../komponenter/Samhandler/SamhandlerTabell';
-import { useSamhandlerRequest } from '../../komponenter/Samhandler/useSamhandler';
+import { SamhandlerTabell } from '../../komponenter/samhandler/SamhandlerTabell';
+import { useSamhandlerRequest } from '../../komponenter/samhandler/useSamhandler';
 import { fagsakStatus, FagsakType } from '../../typer/fagsak';
 import type { ISamhandlerInfo } from '../../typer/samhandler';
 import { formaterIdent } from '../../utils/formatter';
